% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handlers.R
\name{coerce_upload_types}
\alias{coerce_upload_types}
\title{Coerce uploaded columns to match reference types}
\usage{
coerce_upload_types(uploaded, reference)
}
\arguments{
\item{uploaded}{data.frame from upload}

\item{reference}{data.frame with target column types}
}
\value{
data.frame with coerced types
}
\description{
Coerce uploaded columns to match reference types
}
\keyword{internal}
