% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{linked_cells_server}
\alias{linked_cells_server}
\title{LinkedCells Server Module}
\usage{
linked_cells_server(id, data, num_locked_rows = 0, link_fn, tolerance = 0.1)
}
\description{
LinkedCells Server Module
}
