% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handlers.R
\name{read_linked_cells_xlsx}
\alias{read_linked_cells_xlsx}
\title{Read Excel file as data.frame}
\usage{
read_linked_cells_xlsx(file)
}
\arguments{
\item{file}{Path to .xlsx file}
}
\value{
data.frame
}
\description{
Read Excel file as data.frame
}
\keyword{internal}
