% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_utils.R
\name{run_async_compute}
\alias{run_async_compute}
\title{Run computation asynchronously}
\usage{
run_async_compute(expr, on_success, on_error)
}
\description{
Run computation asynchronously
}
\keyword{internal}
