% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handlers.R
\name{write_linked_cells_xlsx}
\alias{write_linked_cells_xlsx}
\title{Write data.frame to Excel}
\usage{
write_linked_cells_xlsx(data, file)
}
\arguments{
\item{data}{data.frame to write}

\item{file}{Output file path}
}
\description{
Write data.frame to Excel
}
\keyword{internal}
