% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{linked_cells_ui}
\alias{linked_cells_ui}
\title{LinkedCells UI Component}
\usage{
linked_cells_ui(id)
}
\arguments{
\item{id}{Module ID (character string)}
}
\description{
Creates the user interface for a linked cells table.
}
